feat: add forecast request
