+++ .env
@@ -1,1 +1,1 @@
-API_KEY = 'a1aac202a64668472fff5a3689bc4c61'
+API_KEY='a1aac202a64668472fff5a3689bc4c61'
